+++ packages/experimental/Checkbox/src/CheckboxTokens.win32.ts
@@ -57,7 +57,8 @@ export const defaultCheckboxTokens: TokenSettings<CheckboxTokens, Theme> = (t: T
       color: t.colors.neutralForeground1,
       checked: {
         checkboxBackgroundColor: t.colors.brandBackgroundHover,
-        checkboxBorderColor: t.colors.brandBackgroundHover,
+        checkboxBorderColor: t.colors.compoundBrandStroke1Hover,
+        checkmarkColor: t.colors.neutralForegroundOnBrandHover,
       },
     },
     pressed: {
@@ -68,12 +69,13 @@ export const defaultCheckboxTokens: TokenSettings<CheckboxTokens, Theme> = (t: T
       color: t.colors.neutralForeground1,
       checked: {
         checkboxBackgroundColor: t.colors.brandBackgroundPressed,
-        checkboxBorderColor: t.colors.brandBackgroundPressed,
+        checkboxBorderColor: t.colors.compoundBrandStroke1Pressed,
+        checkmarkColor: t.colors.neutralForegroundOnBrandPressed,
       },
     },
     checked: {
       checkboxBackgroundColor: t.colors.brandBackground,
-      checkboxBorderColor: t.colors.brandBackground,
+      checkboxBorderColor: t.colors.compoundBrandStroke1Hover,
       checkmarkOpacity: 1,
       checkmarkColor: t.colors.neutralForegroundOnBrand,
       disabled: {