+++ packages/theming/win32-theme/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/win32-theme
 
-This log was last generated on Thu, 03 Mar 2022 20:20:09 GMT and should not be manually modified.
+This log was last generated on Fri, 04 Mar 2022 21:17:46 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.14.13
+
+Fri, 04 Mar 2022 21:17:46 GMT
+
+### Patches
+
+- Update entries (ruaraki@microsoft.com)
 
 ## 0.14.12